refactor(rheem-graphchi): refactor module
* separator platform from plugin
* add aggregators for Mappings and ChannelConversions
