+++ rheem-platforms/rheem-graphchi/src/main/java/org/qcri/rheem/graphchi/channels/ChannelConversions.java
@@ -1,0 +1,19 @@
+package org.qcri.rheem.graphchi.channels;
+
+import org.qcri.rheem.core.optimizer.channels.ChannelConversion;
+import org.qcri.rheem.graphchi.platform.GraphChiPlatform;
+
+import java.util.Collection;
+import java.util.Collections;
+
+/**
+ * {@link ChannelConversion}s for the {@link GraphChiPlatform}.
+ */
+public class ChannelConversions {
+
+    /**
+     * All {@link ChannelConversion}s.
+     */
+    public static Collection<ChannelConversion> ALL = Collections.emptyList();
+
+}
