+++ rheem-platforms/rheem-graphchi/src/main/java/org/qcri/rheem/graphchi/mappings/Mappings.java
@@ -1,0 +1,17 @@
+package org.qcri.rheem.graphchi.mappings;
+
+import org.qcri.rheem.core.mapping.Mapping;
+
+import java.util.Collection;
+import java.util.Collections;
+
+/**
+ * Register for {@link Mapping}s for this platform.
+ */
+public class Mappings {
+
+    public static final Collection<Mapping> ALL = Collections.singletonList(
+            new PageRankMapping()
+    );
+
+}
