+++ rheem-tests/src/test/java/org/qcri/rheem/tests/GraphChiIntegrationIT.java
@@ -3,7 +3,8 @@ package org.qcri.rheem.tests;
 import org.junit.Test;
 import org.qcri.rheem.core.api.RheemContext;
 import org.qcri.rheem.core.plan.rheemplan.RheemPlan;
-import org.qcri.rheem.graphchi.GraphChiPlatform;
+import org.qcri.rheem.graphchi.platform.GraphChiPlatform;
+import org.qcri.rheem.graphchi.plugin.GraphChiPlugin;
 import org.qcri.rheem.java.JavaPlatform;
 import org.qcri.rheem.spark.platform.SparkPlatform;
 
@@ -15,14 +16,14 @@ public class GraphChiIntegrationIT {
     @Test
     public void testPageRankWithJava() {
         RheemPlan rheemPlan = RheemPlans.createCrossCommunityPageRank();
-        RheemContext rc = new RheemContext().with(JavaPlatform.getInstance()).with(GraphChiPlatform.getInstance());
+        RheemContext rc = new RheemContext().with(JavaPlatform.getInstance()).with(new GraphChiPlugin());
         rc.execute(rheemPlan);
     }
 
     @Test
     public void testPageRankWithSpark() {
         RheemPlan rheemPlan = RheemPlans.createCrossCommunityPageRank();
-        RheemContext rc = new RheemContext().with(SparkPlatform.getInstance()).with(GraphChiPlatform.getInstance());
+        RheemContext rc = new RheemContext().with(SparkPlatform.getInstance()).with(new GraphChiPlugin());
         rc.execute(rheemPlan);
     }
